requirements.txt
torch>=1.12
torchvision
transformers>=4.30
sentence-transformers
faiss-cpu
scikit-learn
pandas
numpy
Pillow
tqdm
matplotlib
opencv-python
streamlit
grad-cam  # or pytorch-grad-cam
evaluate
